requirements.txt
Django==4.1.6
django-cors-headers==3.13.0
djangorestframework==3.14.0
requests==2.28.2
openpyxl==3.1.1
